requirements.txt
Django==4.2.1
django-allauth==0.54.0
folium==0.14.0
geocoder==1.38.1
django-crispy-forms==2.0
flake8==5.0.4
flake8-broken-line==0.6.0
flake8-isort==6.0.0
flake8-plugin-utils==1.3.2
flake8-return==1.2.0